bbv: Use RawClient for electrum module
Fortanix SGX does not support using the operating systems resolver. In
order to do name resolution ourselves we need to be able to pass in a
connected TCP stream to `electrum-client`. This is supported by
`electrum-client` without need for changes.

Use `electrum_client::RawClient<TcpStream>` instead of
`electrum_client::Client` to connect to the electrum instance.